Fix int array input to arc_length_to_parameter. Fixes issue 52.

+++ src/splinebox/spline_curves.py
@@ -550,7 +550,7 @@ class Spline:
         if not isinstance(s, np.ndarray):
             s = np.array([s])
         sort_indices = np.argsort(s)
-        results = np.zeros_like(s)
+        results = np.zeros_like(s, dtype=float)
 
         current_value = 0
         lower_bound = 0
